Added a partially working DotnetCore project for Cas. This uses middleware to sign in the user and Cookie Authentication to save the user credentials.
The middleware functions similarly to how the module works in the dotnet project.

Much of the stuff was copied and pasted from the DotNet project. Additionally a parameter for the context and options was added to most of the methods.

I did not do much testing and only tested it for my one use-case (signing into my Cas20 server).