Optional Case Sensitivity
I was running into a problem on one of my projects where the user wouldn't be authenticated because the username inputted into Cas had a different case than the username stored in my database (mobile devices tend to capitalize the first letter). This change will force the username to lower case if it is not case sensitvie so that users don't have to worry about it.